Do not crash if IDs are >2^31 in LAMMPS dump files
id is still truncated to int

+++ src/model/io/LammpsAsciiDumpLoader.java
@@ -264,7 +264,7 @@ public class LammpsAsciiDumpLoader extends MDFileLoader {
 					}
 
 					if (numberColumn != -1)
-						number = Integer.parseInt(parts[numberColumn]);
+						number = (int)Long.parseLong(parts[numberColumn]);
 
 					if (scaledCoords){
 						pos.x = Float.parseFloat(parts[xColumn + 0])*idc.boxSizeX.x - idc.offset.x;
